wrapper that generates a series of ejection inputs for eject3dfun
Script that generates a suite of ejection velocities, block diameters, and ejection directions, then calls the function eject3dfunc, located in modules eject3dfun.  Eject3dfunc returns the x, y, and z landing locations, which are written out in a table to stdout and to an output file.  The range of inputs are specified in parameters within this file.